feat(recipe): add new recipe for vegan tuna
A new recipe has been added to the website, for vegan tuna, a new image
has been added for the tuna and a link to the page was added to the
index page. A small fix in the formatting of the image on the vegan
tacos page is also included.